+++ cmd/main.go
@@ -15,9 +15,9 @@ import (
 	"github.com/naturalselectionlabs/rss3-global-indexer/internal/database/dialer"
 	"github.com/naturalselectionlabs/rss3-global-indexer/internal/hub"
 	"github.com/naturalselectionlabs/rss3-global-indexer/internal/nameresolver"
-	"github.com/naturalselectionlabs/rss3-global-indexer/internal/service/epoch"
 	"github.com/naturalselectionlabs/rss3-global-indexer/internal/service/indexer"
 	"github.com/naturalselectionlabs/rss3-global-indexer/internal/service/scheduler"
+	"github.com/naturalselectionlabs/rss3-global-indexer/internal/service/settler"
 	"github.com/redis/go-redis/v9"
 	"github.com/samber/lo"
 	"github.com/spf13/cobra"
@@ -89,7 +89,7 @@ var command = cobra.Command{
 
 var indexCommand = &cobra.Command{
 	Use: "index",
-	RunE: func(cmd *cobra.Command, args []string) error {
+	RunE: func(cmd *cobra.Command, _ []string) error {
 		flags = cmd.PersistentFlags()
 
 		config, err := config.Setup(lo.Must(flags.GetString(flag.KeyConfig)))
@@ -130,7 +130,7 @@ var indexCommand = &cobra.Command{
 
 var schedulerCommand = &cobra.Command{
 	Use: "scheduler",
-	RunE: func(cmd *cobra.Command, args []string) error {
+	RunE: func(cmd *cobra.Command, _ []string) error {
 		flags = cmd.PersistentFlags()
 
 		config, err := config.Setup(lo.Must(flags.GetString(flag.KeyConfig)))
@@ -169,9 +169,9 @@ var schedulerCommand = &cobra.Command{
 	},
 }
 
-var epochCommand = &cobra.Command{
+var settlerCommand = &cobra.Command{
 	Use: "epoch",
-	RunE: func(cmd *cobra.Command, args []string) error {
+	RunE: func(cmd *cobra.Command, _ []string) error {
 		flags = cmd.PersistentFlags()
 
 		config, err := config.Setup(lo.Must(flags.GetString(flag.KeyConfig)))
@@ -195,7 +195,7 @@ var epochCommand = &cobra.Command{
 
 		redisClient := redis.NewClient(options)
 
-		instance, err := epoch.New(cmd.Context(), databaseClient, redisClient, *config)
+		instance, err := settler.New(cmd.Context(), databaseClient, redisClient, *config)
 		if err != nil {
 			return err
 		}
@@ -245,13 +245,13 @@ func init() {
 
 	command.AddCommand(indexCommand)
 	command.AddCommand(schedulerCommand)
-	command.AddCommand(epochCommand)
+	command.AddCommand(settlerCommand)
 
 	command.PersistentFlags().String(flag.KeyConfig, "./deploy/config.yaml", "config file path")
 	indexCommand.PersistentFlags().String(flag.KeyConfig, "./deploy/config.yaml", "config file path")
 	schedulerCommand.PersistentFlags().String(flag.KeyConfig, "./deploy/config.yaml", "config file path")
 	schedulerCommand.PersistentFlags().String(flag.KeyServer, "detector", "server name")
-	epochCommand.PersistentFlags().String(flag.KeyConfig, "./deploy/config.yaml", "config file path")
+	settlerCommand.PersistentFlags().String(flag.KeyConfig, "./deploy/config.yaml", "config file path")
 }
 
 func main() {
