+++ internal/service/settler/server.go
@@ -1,4 +1,4 @@
-package epoch
+package settler
 
 import (
 	"context"
