+++ internal/service/settler/settler.go
@@ -1,4 +1,4 @@
-package epoch
+package settler
 
 import (
 	"context"
